+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Reflection-Util change log
 
-## 2.0.0 (????-??-??)
+## 2.0.0 (2026-08-05)
 
 - Renamed `ReflectionPlume` to `ReflectionP`.
 - Requires Java 17 to run.